requirements.txt
speechrecognition==3.8.1
wikipedia==1.4.0
googlesearch-python==1.0.0
pyttsx3==2.90
